# db/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Contract SIMS PPOB - MUKLIS APRILIANSYAH",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:3000"
        }
    ],
    "paths": {
...
        "/login": {
            "post": {
                "tags": [
                    "1. Module Membership"
                ],
                "summary": "login",
                "operationId": "loginPOST",
                "description": "API Login Public (Tidak perlu Token untuk mengaksesnya)",
                "parameters": [],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "example": {
                                    "email": "user@nutech-integrasi.com",
                                    "password": "abcdef1234"
                                  }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Request Successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                "type": "object",
                                "example": {
                                    "status": 0,
                                    "message": "Login Sukses",
                                    "data": {
                                      "token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJkYXRhIjoiNTRVdXRjYTdCS0ZPX0ZUZGZ1bXlJem9zSTRKa1VxUGZVZ0ROSTUwelRTQlo2aHoyY0hKZ1VMb1loM09HUUd0ekQxV3dTX194aHBNZTE2SGFscVRzcEhjS21UclJ3S2FYYmZob3AzdzFFUHJ2NFdBQmk1c0RpdV9DSnZTSWt2MDFTbEU0QU5pbVB0bUx5azZoUzlOalVQNEZaVVpfRVBtcEk4Y3pNc3ZWa2JFPSIsImlhdCI6MTYyNjkyODk3MSwiZXhwIjoyNTU2MTE4Nzk4fQ.9C9NvhZYKivhGWnrjo4Wr1Rv-wur1wCm0jqfK9XDD8U"
                                    }
                                  }
                            }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                "type": "object",
                                "example": {
                                    "status": 102,
                                    "message": "Parameter email tidak sesuai format",
                                    "data": null
                                  }
                            }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                "type": "object",
                                "example": {
                                    "status": 103,
                                    "message": "Username atau password salah",
                                    "data": null
                                  }
                            }
                            }
                        }
                    }
                },
                "x-swagger-router-controller": "ModuleMembership"
            }
        },
        "/profile": {
            "get": {
                "tags": [
                    "1. Module Membership"
                ],
                "summary": "profile",
                "operationId": "profileGET",
                "description": "API Profile Private (memerlukan Token untuk mengaksesnya)",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "Request Successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                "type": "object",
                                "example": {
                                    "status": 0,
                                    "message": "Sukses",
                                    "data": {
                                      "email": "user@nutech-integrasi.com",
                                      "first_name": "User",
                                      "last_name": "Nutech",
                                      "profile_image": "https://yoururlapi.com/profile.jpeg"
                                    }
                                  }
                            }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                "type": "object",
                                "example": {
                                    "status": 108,
                                    "message": "Token tidak valid atau kadaluwarsa",
                                    "data": null
                                  }
                            }
                            }
                        }
                    }
                },
                "x-swagger-router-controller": "ModuleMembership"
            }
        },
        "/profile/update": {
            "put": {
                "tags": [
                    "1. Module Membership"
                ],
                "summary": "update",
                "operationId": "updatePUT",
                "description": "API Update Profile Private (memerlukan Token untuk mengaksesnya)",
                "parameters": [],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "example": {
                                    "first_name": "User Edited",
                                    "last_name": "Nutech Edited"
                                  }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Request Successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                "type": "object",
                                "example": {
                                    "status": 0,
                                    "message": "Update Pofile berhasil",
                                    "data": {
                                      "email": "user@nutech-integrasi.com",
                                      "first_name": "User Edited",
                                      "last_name": "Nutech Edited",
                                      "profile_image": "https://yoururlapi.com/profile.jpeg"
                                    }
                                  }
                            }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                "type": "object",
                                "example": {
                                    "status": 108,
                                    "message": "Token tidak valid atau kadaluwarsa",
                                    "data": null
                                  }
                            }
                            }
                        }
                    }
                },
                "x-swagger-router-controller": "ModuleMembership"
            }
        },
        "/profile/image": {
            "put": {
                "tags": [
                    "1. Module Membership"
                ],
                "summary": "image",
                "operationId": "imagePUT",
                "description": "API Upload Profile Image Private (memerlukan Token untuk mengaksesnya)",
                "parameters": [],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                        "schema": {
                            "type": "object",
                            "properties": {
                            "file": {
                                "type": "string",
                                "format": "binary",
                                "description": "The file to be uploaded"
                            }
                            },
                            "required": ["file"]
                        }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Request Successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                "type": "object",
                                "example": {
                                    "status": 0,
                                    "message": "Update Profile Image berhasil",
                                    "data": {
                                      "email": "user@nutech-integrasi.com",
                                      "first_name": "User Edited",
                                      "last_name": "Nutech Edited",
                                      "profile_image": "https://yoururlapi.com/profile-updated.jpeg"
                                    }
                                  }
                            }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                "type": "object",
                                "example": {
                                    "status": 102,
                                    "message": "Format Image tidak sesuai",
                                    "data": null
                                  }
                            }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                "type": "object",
                                "example": {
                                    "status": 108,
                                    "message": "Token tidak valid atau kadaluwarsa",
                                    "data": null
                                  }
                            }
                            }
                        }
                    }
                },
                "x-swagger-router-controller": "ModuleMembership"
            }
        },
        "/banner": {
            "get": {
                "tags": [
                    "2. Module Information"
                ],
                "summary": "banner",
                "operationId": "bannerGET",
                "description": "API Banner Private (memerlukan Token untuk mengaksesnya)",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "Request Successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                "type": "object",
                                "example": {
                                    "status": 0,
                                    "message": "Sukses",
                                    "data": [
                                      {
                                        "banner_name": "Banner 1",
                                        "banner_image": "https://nutech-integrasi.app/dummy.jpg",
                                        "description": "Lerem Ipsum Dolor sit amet"
                                      },
                                      {
                                        "banner_name": "Banner 2",
                                        "banner_image": "https://nutech-integrasi.app/dummy.jpg",
                                        "description": "Lerem Ipsum Dolor sit amet"
                                      },
                                      {
                                        "banner_name": "Banner 3",
                                        "banner_image": "https://nutech-integrasi.app/dummy.jpg",
                                        "description": "Lerem Ipsum Dolor sit amet"
                                      },
                                      {
                                        "banner_name": "Banner 4",
                                        "banner_image": "https://nutech-integrasi.app/dummy.jpg",
                                        "description": "Lerem Ipsum Dolor sit amet"
                                      },
                                      {
                                        "banner_name": "Banner 5",
                                        "banner_image": "https://nutech-integrasi.app/dummy.jpg",
                                        "description": "Lerem Ipsum Dolor sit amet"
                                      },
                                      {
                                        "banner_name": "Banner 6",
                                        "banner_image": "https://nutech-integrasi.app/dummy.jpg",
                                        "description": "Lerem Ipsum Dolor sit amet"
                                      }
                                    ]
                                  }
                            }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                "type": "object",
                                "example": {
                                    "status": 108,
                                    "message": "Token tidak valid atau kadaluwarsa",
                                    "data": null
                                  }
                            }
                            }
                        }
                    }
                },
                "x-swagger-router-controller": "ModuleMembership"
            }
        },
        "/services": {
            "get": {
                "tags": [
                    "2. Module Information"
                ],
                "summary": "services",
                "operationId": "servicesGET",
                "description": "API Services Private (memerlukan Token untuk mengaksesnya)",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "Request Successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                "type": "object",
                                "example": {
                                    "status": 0,
                                    "message": "Sukses",
                                    "data": [
                                      {
                                        "service_code": "PAJAK",
                                        "service_name": "Pajak PBB",
                                        "service_icon": "https://nutech-integrasi.app/dummy.jpg",
                                        "service_tariff": 40000
                                      },
                                      {
                                        "service_code": "PLN",
                                        "service_name": "Listrik",
                                        "service_icon": "https://nutech-integrasi.app/dummy.jpg",
                                        "service_tariff": 10000
                                      },
                                      {
                                        "service_code": "PDAM",
                                        "service_name": "PDAM Berlangganan",
                                        "service_icon": "https://nutech-integrasi.app/dummy.jpg",
                                        "service_tariff": 40000
                                      },
                                      {
                                        "service_code": "PULSA",
                                        "service_name": "Pulsa",
                                        "service_icon": "https://nutech-integrasi.app/dummy.jpg",
                                        "service_tariff": 40000
                                      },
                                      {
                                        "service_code": "PGN",
                                        "service_name": "PGN Berlangganan",
                                        "service_icon": "https://nutech-integrasi.app/dummy.jpg",
                                        "service_tariff": 50000
                                      },
                                      {
                                        "service_code": "MUSIK",
                                        "service_name": "Musik Berlangganan",
                                        "service_icon": "https://nutech-integrasi.app/dummy.jpg",
                                        "service_tariff": 50000
                                      },
                                      {
                                        "service_code": "TV",
                                        "service_name": "TV Berlangganan",
                                        "service_icon": "https://nutech-integrasi.app/dummy.jpg",
                                        "service_tariff": 50000
                                      },
                                      {
                                        "service_code": "PAKET_DATA",
                                        "service_name": "Paket data",
                                        "service_icon": "https://nutech-integrasi.app/dummy.jpg",
                                        "service_tariff": 50000
                                      },
                                      {
                                        "service_code": "VOUCHER_GAME",
                                        "service_name": "Voucher Game",
                                        "service_icon": "https://nutech-integrasi.app/dummy.jpg",
                                        "service_tariff": 100000
                                      },
                                      {
                                        "service_code": "VOUCHER_MAKANAN",
                                        "service_name": "Voucher Makanan",
                                        "service_icon": "https://nutech-integrasi.app/dummy.jpg",
                                        "service_tariff": 100000
                                      },
                                      {
                                        "service_code": "QURBAN",
                                        "service_name": "Qurban",
                                        "service_icon": "https://nutech-integrasi.app/dummy.jpg",
                                        "service_tariff": 200000
                                      },
                                      {
                                        "service_code": "ZAKAT",
                                        "service_name": "Zakat",
                                        "service_icon": "https://nutech-integrasi.app/dummy.jpg",
                                        "service_tariff": 300000
                                      }
                                    ]
                                  }
                            }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                "type": "object",
                                "example": {
                                    "status": 108,
                                    "message": "Token tidak valid atau kadaluwarsa",
                                    "data": null
                                  }
                            }
                            }
                        }
                    }
                },
                "x-swagger-router-controller": "ModuleMembership"
            }
        },
        "/balance": {
           "get": {
                "tags": [
                    "3. Module Transaction"
                ],
                "summary": "balance",
                "operationId": "balanceGET",
                "description": "API Balance Private (memerlukan Token untuk mengaksesnya)",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "Request Successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                "type": "object",
                                "example": {
                                    "status": 0,
                                    "message": "Get Balance Berhasil",
                                    "data": {
                                      "balance": 1000000
                                    }
                                  }
                            }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                "type": "object",
                                "example": {
                                    "status": 108,
                                    "message": "Token tidak valid atau kadaluwarsa",
                                    "data": null
                                  }
                            }
                            }
                        }
                    }
                },
                "x-swagger-router-controller": "ModuleMembership"
            }
        },
        "/topup": {
            "post": {
                "tags": [
                    "3. Module Transaction"
                ],
                "summary": "topup",
                "operationId": "topupPOST",
                "description": "API Topup Private (memerlukan Token untuk mengaksesnya)",
                "parameters": [],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "example": {
                                    "top_up_amount": 1000000
                                  }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Request Successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                "type": "object",
                                "example": {
                                    "status": 0,
                                    "message": "Top Up Balance berhasil",
                                    "data": {
                                      "balance": 2000000
                                    }
                                  }
                            }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                "type": "object",
                                "example": {
                                    "status": 102,
                                    "message": "Parameter amount hanya boleh angka dan tidak boleh lebih kecil dari 0",
                                    "data": null
                                  }
                            }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                "type": "object",
                                "example": {
                                    "status": 108,
                                    "message": "Token tidak valid atau kadaluwarsa",
                                    "data": null
                                  }
                            }
                            }
                        }
                    }
                },
                "x-swagger-router-controller": "ModuleMembership"
            }
        },
        "/transaction": {
           "post": {
                "tags": [
                    "3. Module Transaction"
                ],
                "summary": "transaction",
                "operationId": "transactionPOST",
                "description": "API Transaction Private (memerlukan Token untuk mengaksesnya)",
                "parameters": [],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "example": {
                                    "service_code": "PULSA"
                                  }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Request Successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                "type": "object",
                                "example": {
                                    "status": 0,
                                    "message": "Transaksi berhasil",
                                    "data": {
                                      "invoice_number": "INV17082023-001",
                                      "service_code": "PLN_PRABAYAR",
                                      "service_name": "PLN Prabayar",
                                      "transaction_type": "PAYMENT",
                                      "total_amount": 10000,
                                      "created_on": "2023-08-17T10:10:10.000Z"
                                    }
                                  }
                            }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                "type": "object",
                                "example": {
                                    "status": 102,
                                    "message": "Service ataus Layanan tidak ditemukan",
                                    "data": null
                                  }
                            }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                "type": "object",
                                "example": {
                                    "status": 108,
                                    "message": "Token tidak valid atau kadaluwarsa",
                                    "data": null
                                  }
                            }
                            }
                        }
                    }
                },
                "x-swagger-router-controller": "ModuleMembership"
            }
        },
...
    },
...
}